+++ src/db/initMongoConnection.js
@@ -6,11 +6,11 @@ export const initMongoConnection = async () => {
     const user = getEnvVar('MONGODB_USER');
     const pwd = getEnvVar('MONGODB_PASSWORD');
     const url = getEnvVar('MONGODB_URL');
-    const db = getEnvVar('MONGODB_DB');
+    // const db = getEnvVar('MONGODB_DB');
 
     await mongoose.connect(
-      `mongodb+srv://${user}:${pwd}@${url}/${db}?retryWrites=true&w=majority`,
-      // `mongodb+srv://${user}:${pwd}@${url}/?retryWrites=true&w=majority&appName=SportPoint`,
+      // `mongodb+srv://${user}:${pwd}@${url}/${db}?retryWrites=true&w=majority`,
+      `mongodb+srv://${user}:${pwd}@${url}/?retryWrites=true&w=majority&appName=SportPoint`,
     );
 
     console.log('Mongo connection successfully established!');
